move modules, schema and defaults to src
