+++ MANIFEST
@@ -1,0 +1,16 @@
+# file GENERATED by distutils, do NOT edit
+MANIFEST.in
+README.md
+setup.py
+__init__.py
+bin/append_ctable.py
+bin/dswx_hls.py
+bin/dswx_hls_download.py
+src/extern/__init__.py
+src/extern/validate_cloud_optimized_geotiff.py
+src/extern/__pycache__/__init__.cpython-37.pyc
+src/extern/__pycache__/__init__.cpython-39.pyc
+src/extern/__pycache__/validate_cloud_optimized_geotiff.cpython-37.pyc
+src/extern/__pycache__/validate_cloud_optimized_geotiff.cpython-39.pyc
+src/schemas/dswx_hls.yaml
+src/defaults/dswx_hls.yaml
