+++ MANIFEST.in
@@ -1,5 +1,5 @@
 include MANIFEST.in
 include *.md
-recursive-include defaults *
-recursive-include extern *
-recursive-include schemas *
+recursive-include src/defaults *
+recursive-include src/schemas *
+recursive-include src/extern *
