+++ __init__.py
@@ -1,0 +1,2 @@
+import .bin
+import .src
