--- modules/__init__.py
@@ -1,1 +1,0 @@
-from dswx_hls import *
