+++ src/extern/validate_cloud_optimized_geotiff.py
